% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr_matrix2df.R
\name{dtr_matrix2df}
\alias{dtr_matrix2df}
\title{dtr_matrix2df}
\usage{
dtr_matrix2df(matrix, cat_vars = NULL)
}
\arguments{
\item{matrix}{An R `matrix` object.}

\item{cat_vars}{A character vector with colnames that should be converted to
`factor` (default: NULL).}
}
\value{
A `data.table` is returned.
}
\description{
Data transformation: Converts a `matrix` to `data.table` and
  encodes categorical variables as `factor`.
}
\examples{
data("iris")
mat <- data.matrix(iris)
dataset <- dtr_matrix2df(mat)
str(dataset)

dataset <- dtr_matrix2df(mat, cat_vars = "Species")
str(dataset)
}
