% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_icolnames.R
\name{icolnames}
\alias{icolnames}
\title{icolnames}
\usage{
icolnames(df)
}
\arguments{
\item{df}{A data.frame object.}
}
\value{
A \code{data.table} with the two columns \code{index} and \code{name} is returned.
}
\description{
Return colnames in a table with index numbers.
}
\examples{
data("iris")
icolnames(iris)
}
