% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_append.R
\name{list.append}
\alias{list.append}
\title{list.append}
\usage{
list.append(main_list, append_list)
}
\arguments{
\item{main_list}{A list, to which another should be appended.}

\item{append_list}{A list to append to `main_list` (can be `NULL` or empty).}
}
\value{
A `list` is returned.
}
\description{
Helper function to append an R list.
}
\details{
This function is a save wrapper around combining lists with `c()`
  as it checks for the input types and only appends the new list if its
  length is greater than 0.
}
\examples{
l1 <- list("a" = 1, "b" = 2)
l2 <- list("c" = 3, "d" = 4)
list.append(l1, l2)
}
