% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_argument_catcher.R
\name{misc_argument_catcher}
\alias{misc_argument_catcher}
\title{misc_argument_catcher}
\usage{
misc_argument_catcher(...)
}
\arguments{
\item{...}{Named arguments passed to a function.}
}
\value{
A \code{list} is returned.
}
\description{
Miscellaneous helper function to type-save catch arguments
passed with R's ellipsis ("...").
}
\details{
This function aims at catching arguments that have been passed to an
R function using R's ellipsis ("..."). Its purpos is to catch these
arguments even in the case, if a list with arguments was provided to the
ellipsis.
}
\examples{
misc_argument_catcher(a = 1)
misc_argument_catcher(a = 1, b = 2, c = 3, d = "car")
misc_argument_catcher(list(a = 1, b = 2, c = 3, d = "car"))
misc_argument_catcher(list(a = 1, b = 2, c = 3, d = "car"), f = 9)
}
