% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_recursive_copy.R
\name{misc_recursive_copy}
\alias{misc_recursive_copy}
\title{misc_recursive_copy}
\usage{
misc_recursive_copy(source_dir, target_dir)
}
\arguments{
\item{source_dir}{A character string. The path to the directory to be copied.}

\item{target_dir}{A character string. The target path.}
}
\value{
This function has no return value.
}
\description{
Recursively copying directories and subdirectories.
}
\examples{
\donotrun{
  misc_recursive_copy("/tmp/folder", "/tmp/new_folder/")
}
}
