% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlh_format_xy.R
\name{mlh_format_xy}
\alias{mlh_format_xy}
\title{mlh_format_xy}
\usage{
mlh_format_xy(object, ids)
}
\arguments{
\item{object}{A vector or a data matrix.}

\item{ids}{An integer vector specifying the indices that should be selected
from the object.}
}
\value{
Returns a subset of object.
}
\description{
Machine learning helper function to select a subset from a
data matrix or a response vector.
}
\examples{
data("iris")
mlh_format_xy(iris, c(1:30))
mlh_format_xy(iris[, 5], c(1:30))
}
