% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlh_subset.R
\name{mlh_subset}
\alias{mlh_subset}
\title{mlh_subset}
\usage{
mlh_subset(object, ids)
}
\arguments{
\item{object}{A vector or a data matrix. Supports also subsetting of "Surv"
objects.}

\item{ids}{An integer vector specifying the indices that should be selected
from the object.}
}
\value{
Returns the specified subset of the object.
}
\description{
Machine learning helper function to select a subset from a
data matrix or a response vector.
}
\examples{
data("iris")
mlh_subset(iris, c(1:30))
mlh_subset(iris[, 5], c(1:30))
}
