% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pch_check_available_cores.R
\name{pch_check_available_cores}
\alias{pch_check_available_cores}
\title{pch_check_available_cores}
\usage{
pch_check_available_cores(ncores = -1L)
}
\arguments{
\item{ncores}{An integer. A number of cores requested for parallel computing
(default: `-1L`).}
}
\value{
The function returns an integer that indicates the number of cores
  available. If `ncores <= parallel::detectCores()` the function returns
  `ncores`. If `ncores > parallel::detectCores()`, the function returns
  `parallel::detectCores() - 1L`.
}
\description{
Parallel computing helper function to check for the available
  cores.
}
\examples{
pch_check_available_cores(2)
}
