% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pch_register_parallel.R
\name{pch_register_parallel}
\alias{pch_register_parallel}
\title{pch_register_parallel}
\usage{
pch_register_parallel(ncores)
}
\arguments{
\item{ncores}{An integer. A number of cores requested for parallel computing
(default: `-1L`).}
}
\value{
The function returns a object of class `c("SOCKcluster", "cluster")`.
}
\description{
Parallel computing helper function to register a parallel
  backend.
}
\examples{
cl <- pch_register_parallel(pch_check_available_cores(2))
pch_clean_up(cl)
}
\seealso{
[parallel::makePSOCKcluster()], [doParallel::registerDoParallel()]
}
