% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_mean_sd.R
\name{rep_mean_sd}
\alias{rep_mean_sd}
\title{rep_mean_sd}
\usage{
rep_mean_sd(
  x,
  digits = 2,
  na.rm = TRUE,
  sd_brackets = c("round", "square"),
  sd_prefix = TRUE,
  weighted = FALSE,
  weights = NA
)
}
\description{
rep_mean_sd
}
\examples{
set.seed(123)
x <- rnorm(1000)
rep_mean_sd(x)
rep_mean_sd(rep(1, 10))
rep_mean_sd(x, sd_brackets = "square")
rep_mean_sd(x, sd_brackets = "square", sd_prefix = FALSE)

}
