% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_median_ci.R
\name{rep_median_ci}
\alias{rep_median_ci}
\title{rep_median_ci}
\usage{
rep_median_ci(
  x,
  conf_int,
  digits = 2,
  na.rm = TRUE,
  collapse = "to",
  iqr_brackets = c("round", "square"),
  iqr_prefix = TRUE,
  weighted = FALSE,
  weights = NA
)
}
\description{
rep_median_ci
}
\examples{
set.seed(123)
x <- rnorm(1000)
rep_median_ci(x, conf_int = 95)
rep_median_ci(rep(1, 10), conf_int = 95)
rep_median_ci(x, conf_int = 95, collapse = "-")
rep_median_ci(x, iqr_brackets = "square", conf_int = 50)

}
