% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_median_iqr.R
\name{rep_median_iqr}
\alias{rep_median_iqr}
\title{rep_median_iqr}
\usage{
rep_median_iqr(
  x,
  digits = 2,
  na.rm = TRUE,
  collapse = "to",
  iqr_brackets = c("round", "square"),
  iqr_prefix = TRUE
)
}
\description{
rep_median_iqr
}
\examples{
set.seed(123)
x <- rnorm(1000)
rep_median_iqr(x)
rep_median_iqr(rep(1, 10))
rep_median_iqr(x, collapse = "-")
rep_median_iqr(x, iqr_brackets = "square")
rep_median_iqr(x, iqr_brackets = "square", iqr_prefix = FALSE)
rep_median_iqr(x, collapse = ";", iqr_prefix = FALSE)

}
