% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_pval.R
\name{rep_pval}
\alias{rep_pval}
\title{rep_pval}
\usage{
rep_pval(p, threshold = 0.001, digits = 3L)
}
\arguments{
\item{p}{The p-value that should be formatted.}

\item{threshold}{A threshold to indicate that only "< threshold" is
printed as output (default: 0.001).}

\item{digits}{The number of digits of the formatted p-value (digits).}
}
\value{
A character with the formatted p-value.
}
\description{
Reporting helper function: formats p-value.
}
\details{
If the p-value is lower than the threshold, the output of the
function is "< threshold". Otherwise, the p-value is formatted to the
number of digits.
}
\examples{
rep_pval(0.032)
rep_pval(0.00032)

}
