% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_sum_pct.R
\name{rep_sum_pct}
\alias{rep_sum_pct}
\title{rep_sum_pct}
\usage{
rep_sum_pct(
  count,
  count_reference,
  digits = 2,
  na.rm = TRUE,
  brackets = c("round", "square"),
  suffix = TRUE
)
}
\description{
rep_sum_pct
}
\examples{
rep_sum_pct(count = 40, count_reference = 200)
rep_sum_pct(count = 40, count_reference = 200, brackets = "square")
rep_sum_pct(count = 40, count_reference = 200, brackets = "square", suffix = FALSE)

}
