% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts_normalize.R
\name{sts_normalize}
\alias{sts_normalize}
\title{sts_normalize}
\usage{
sts_normalize(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector of type `numeric`.}

\item{na.rm}{A logical to indicate, if missings should be removed.}
}
\value{
Returns a vector of same length as `x` with values normalized between
  zero and one. If `x` contains missings and `na.rm = TRUE`, the missings are
  removed before normalization; otherwise, a vector of `NA` is returend.
}
\description{
Statistic helper function to normalize a continuous variable
  between zero and one.
}
\examples{
sts_normalize(1:100)
}
