% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_duplicated_by_names.R
\name{misc_duplicated_by_names}
\alias{misc_duplicated_by_names}
\title{misc_duplicated_by_names}
\usage{
misc_duplicated_by_names(object, ...)
}
\arguments{
\item{object}{An R object that has names.}

\item{...}{Named arguments passed on to \code{duplicated}.}
}
\value{
Returns a logical vector of \code{length(object)} with \code{TRUE} indicating
the identified items with duplicated names.
}
\description{
Miscellaneous helper function to detect items in an object
with duplicated names, e.g. in named vectors or named lists.
}
\examples{
misc_duplicated_by_names(list(a = 1, a = 1))
}
\seealso{
\code{\link[base:duplicated]{base::duplicated()}}
}
