% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_subset_options.R
\name{misc_subset_options}
\alias{misc_subset_options}
\title{misc_subset_options}
\usage{
misc_subset_options(keyword)
}
\arguments{
\item{keyword}{A character. The keyword to subset the R options.}
}
\value{
A \code{list} is returned, containing the subset of R's \code{options()} that
matches with the \code{keyword}.
}
\description{
Miscellaneous helper function to subset R options by a keyword.
}
\details{
This function subsets R's \code{options()} by a keyword. It returns a
list of all available options that match with the \code{keyword}. The \code{keyword}
is evaluated as a regular expression.
}
\examples{
misc_subset_options("default")
}
