% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pch_clean_up.R
\name{pch_clean_up}
\alias{pch_clean_up}
\title{pch_clean_up}
\usage{
pch_clean_up(cl)
}
\arguments{
\item{cl}{A cluster object of class \code{c("SOCKcluster", "cluster")}.}
}
\value{
The function returns nothing. Internally, it calls
\code{parallel::stopCluster()} and \code{foreach::registerDoSEQ()}.
}
\description{
Parallel computing helper function to clean up the parallel
backend.
}
\examples{
if (require("doParallel") && require("foreach")) {
  cl <- pch_register_parallel(pch_check_available_cores(2))
  pch_clean_up(cl)
}
}
\seealso{
\code{\link[parallel:makeCluster]{parallel::stopCluster()}}, \code{\link[foreach:registerDoSEQ]{foreach::registerDoSEQ()}}
}
