% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pch_register_parallel.R
\name{pch_register_parallel}
\alias{pch_register_parallel}
\title{pch_register_parallel}
\usage{
pch_register_parallel(ncores)
}
\arguments{
\item{ncores}{An integer. A number of cores requested for parallel computing
(default: \code{-1L}).}
}
\value{
The function returns a object of class \code{c("SOCKcluster", "cluster")},
created with \code{parallel::makePSOCKcluster()}.
}
\description{
Parallel computing helper function to register a parallel
backend.
}
\examples{
if (require("doParallel") && require("foreach")) {
  cl <- pch_register_parallel(pch_check_available_cores(2))
  pch_clean_up(cl)
}
}
\seealso{
\code{\link[parallel:makeCluster]{parallel::makePSOCKcluster()}}, \code{\link[doParallel:registerDoParallel]{doParallel::registerDoParallel()}}
}
