% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_parallel_coordinates.R
\name{plt_parallel_coordinates}
\alias{plt_parallel_coordinates}
\title{plt_parallel_coordinates}
\usage{
plt_parallel_coordinates(
  data,
  cols = NULL,
  color_variable = NULL,
  color_args = list(alpha = 0.6, begin = 0.1, end = 0.9, option = "inferno", direction =
    1),
  line_jitter = list(w = 0.04, h = 0.04),
  text_label_size = 3.5
)
}
\arguments{
\item{data}{A data.table object with the columns containing the parameters to
be plotted with the parallel coordinates plot.}

\item{cols}{A character vector with column names to subset `data` (default:
`NULL`).}

\item{color_variable}{A character. The name of the column to be used to
colorize the lines of the plot (default: `NULL`).}

\item{color_args}{A list with parameters for the color gradient (see
details).}

\item{line_jitter}{A list with the elements `w` and `h` to define a line
jitter (default: `list(w = 0.04, h = 0.04)`), which are passed further on
to define the position of `ggplot2::geom_line()`.}

\item{text_label_size}{A numeric value to define the size of the text
annotations (default: `3.5`).}
}
\value{
Returns a parallel coordinates plot as `ggplot2` object.
}
\description{
Parallel coordinates plot
}
\details{
The color gradient of the plotted lines can be defined with a list
  provided to the argument `color_args`. Its default values are
  `alpha = 0.6`, `begin = .1`, `end = .9`, `option = "inferno"`, and
  `direction = 1` and are passed furhter on to
  `ggplot2::scale_color_viridis_c()`.
  The implementation to display categorical variables is still experimental.
}
\examples{
if (require("ggplot2")) {
  data("iris")
  plt_parallel_coordinates(data.table::as.data.table(iris[, -5]))
}
}
\seealso{
[ggplot2::scale_color_viridis_c()]
}
